requirements.txt
asgiref==3.5.0
azt==0.0.7
cffi==1.15.0
click==8.0.4
cryptography==36.0.1
Cython==0.29.28
Django==3.2
django-cors-headers==3.11.0
django-rest-knox==4.2.0
django-restframework==0.0.1
djangorestframework==3.13.1
gunicorn==20.1.0
importlib-metadata==4.11.3
install==1.3.5
mysqlclient==2.1.0
pycparser==2.21
pyfiglet==0.8.post1
pytz==2021.3
pyzt==1.0.1
sqlparse==0.4.2
termcolor==1.1.0
typing-extensions==4.1.1
zipp==3.7.0
